Enhance project configuration and remove Win32 support
Updated `Serin.csproj` to support `AnyCPU` and `x64` platforms, and added application icon `Serin.ico`. Included new projects "Serin App" and "Serin Installer" in the solution with updated configurations for `ARM64` and `x64`.

Modified `Serin-Reader.vcxproj` to remove `Win32` configurations, focusing exclusively on `x64` for both Debug and Release builds, improving performance and compatibility with modern systems.
Note: the installer is on another project, trying to fix git so it pushes to the Serin repo (project for tomorrow).